+++ crates/ext_parquet/src/metadata/page_index/index.rs
@@ -56,13 +56,13 @@ impl<T> PageIndex<T> {
     }
 }
 
-#[derive(Debug, Clone, PartialEq)]
-#[allow(non_camel_case_types)]
 /// Typed statistics for a data page in a column chunk.
 ///
 /// This structure is part of the "Page Index" and is optionally part of
 /// [ColumnIndex] in the parquet file and can be used to skip decoding pages
 /// while reading the file data.
+#[derive(Debug, Clone, PartialEq)]
+#[allow(non_camel_case_types)]
 pub enum Index {
     /// Sometimes reading page index from parquet file
     /// will only return pageLocations without min_max index,
