+++ crates/ext_parquet/src/page.rs
@@ -2,8 +2,8 @@
 
 use crate::basic::{Encoding, Type};
 use crate::errors::{ParquetError, ParquetResult, nyi_err};
-use crate::file::statistics::{self, Statistics};
 use crate::format;
+use crate::metadata::statistics::{self, Statistics};
 
 #[derive(Debug, PartialEq)]
 pub struct DataPageHeader {
